Update freemarker bp to retrieve content and add pagination

+++ scripts/classes/org/craftercms/sites/wordify/SearchHelper.groovy
@@ -114,7 +114,7 @@ class SearchHelper {
       def pagination = [:]
       def posts = searchPosts(categories, page * postsPerPage, postsPerPage, exclude, tags)
 
-      pagination.totalPosts = posts.total instanceof String ? posts.total : posts.total.value
+      pagination.totalPosts = posts.total instanceof String ? posts.total : posts.total.value.toDouble()
       pagination.pages = Math.ceil(pagination.totalPosts/postsPerPage)
       pagination.currentPage = page + 1
